refactor notification: 폴더 정리
