+++ src/main/java/com/anyang/maruni/domain/notification/infrastructure/MockNotificationRecord.java
@@ -2,7 +2,7 @@ package com.anyang.maruni.domain.notification.infrastructure;
 
 import java.time.LocalDateTime;
 
-import com.anyang.maruni.domain.notification.domain.service.NotificationChannelType;
+import com.anyang.maruni.domain.notification.domain.vo.NotificationChannelType;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
