+++ src/main/java/com/anyang/maruni/domain/notification/infrastructure/NotificationHistoryDecorator.java
@@ -1,7 +1,7 @@
 package com.anyang.maruni.domain.notification.infrastructure;
 
 import com.anyang.maruni.domain.notification.domain.entity.NotificationHistory;
-import com.anyang.maruni.domain.notification.domain.service.NotificationChannelType;
+import com.anyang.maruni.domain.notification.domain.vo.NotificationChannelType;
 import com.anyang.maruni.domain.notification.domain.service.NotificationHistoryService;
 import com.anyang.maruni.domain.notification.domain.service.NotificationService;
 import lombok.RequiredArgsConstructor;
