+++ src/test/java/com/anyang/maruni/domain/notification/infrastructure/NotificationHistoryServiceImplTest.java
@@ -2,8 +2,8 @@ package com.anyang.maruni.domain.notification.infrastructure;
 
 import com.anyang.maruni.domain.notification.domain.entity.NotificationHistory;
 import com.anyang.maruni.domain.notification.domain.repository.NotificationHistoryRepository;
-import com.anyang.maruni.domain.notification.domain.service.NotificationChannelType;
-import com.anyang.maruni.domain.notification.domain.service.NotificationStatistics;
+import com.anyang.maruni.domain.notification.domain.vo.NotificationChannelType;
+import com.anyang.maruni.domain.notification.domain.vo.NotificationStatistics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
